requirements.txt
Flask==3.0.3
python-telegram-bot==21.6
APScheduler==3.10.4
gunicorn==22.0.0